requirements.txt
openai
streamlit
docx2txt
pdfminer.six
docx
python-docx